requirements.txt
# 필요한 라이브러리 리스트

fastapi	FastAPI 웹 서버 프레임워크
uvicorn	FastAPI 실행용 ASGI 서버 (uvicorn main:app --reload)
pydantic	요청 데이터 검증 및 모델 정의 (BaseModel)
python-dotenv	.env 파일에서 환경변수 로드 (load_dotenv())
openai	OpenAI GPT API 호출 (OpenAI)